organize report card related files
